+++ lib/dao/transaction_dao.dart
@@ -1,5 +1,5 @@
-import 'transaction_model.dart';
-import 'database_helper.dart';
+import '../models/transaction_model.dart';
+import '../database_helper.dart';
 
 class TransactionDao {
   final dbHelper = DatabaseHelper.instance;
