+++ lib/models/bank_model.dart
