+++ lib/pages/DatabaseViewerPage.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'database_helper.dart';
+import '../database_helper.dart';
 
 class DatabaseViewerPage extends StatefulWidget {
   const DatabaseViewerPage({super.key});
