+++ lib/pages/register.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'bank_dao.dart';
-import 'bank_model.dart';
-import 'user_dao.dart';
-import 'user_model.dart';
-import 'account_dao.dart';
-import 'account_model.dart';
+import '../dao/bank_dao.dart';
+import '../models/bank_model.dart';
+import '../dao/user_dao.dart';
+import '../models/user_model.dart';
+import '../dao/account_dao.dart';
+import '../models/account_model.dart';
 import 'home.dart';
-import 'read_sms.dart';
+import '../utils/read_sms.dart';
 
 class RegisterPage extends StatefulWidget {
   const RegisterPage({super.key});
