+++ lib/pages/splash_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'user_dao.dart';
+import '../dao/user_dao.dart';
 import 'home.dart';
 import 'register.dart';
